[WEEK6] 입력 검증 로직 추가

+++ week6/CoffeeShop/src/main/java/dev/bank/formatter/NumberFormatter.java
@@ -1,7 +1,7 @@
 package dev.bank.formatter;
 
 public class NumberFormatter {
-	public static String formatNumber(int number) {
+	public static String formatNumber(long number) {
 		return String.format("%,d", number);
 	}
 }
